进行整体结构性优化，主要是网络指令处理部分做了优化

+++ plotwidget.h
@@ -47,7 +47,7 @@ public slots:
     void slotPlotClick(QCPAbstractPlottable *plottable, int dataIndex, QMouseEvent *event);
     void slotBeforeReplot();
     void slotRestorePlot(QMouseEvent*);
-    void slotUpdateEnTimeWidth(int* timeWidth);
+    void slotUpdateEnTimeWidth(unsigned short* timeWidth);
 
     //批量刷新
     void slotUpdatePlotDatas(SingleSpectrum, vector<CoincidenceResult>, int refreshTime, int coolTime);
